+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Jigsass Tools Maps Changelog
 
+## v1.0.4
+`jigsass-get`: For behavior that more closely resembles `map-get`, return `null` 
+when a key in the keys chain, which isn't the last key, does not exist or 
+isn't a map.
+
+## v1.0.3
+Update dependencies.
+
 ## v1.0.2
 Fixes a bug where trying to deep-merge with or into an
 empty list would throw an error because a list is not a map.